scale_down/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
cryptography==3.4.7
datastore-entity==0.2.0
Django==3.2.3
Flask==2.0.0
Flask-Login==0.5.0
google==3.0.0
google-api-core==1.27.0
google-auth==1.30.0
google-cloud==0.34.0
google-cloud-core==1.6.0
google-cloud-datastore==2.1.2
google-cloud-storage==1.38.0
google-cloud-vision==2.3.1
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
idna==2.10
image==1.5.33
itsdangerous==2.0.1
Jinja2==3.0.0
libcst==0.3.19
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.9
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
Werkzeug==2.0.1
